// packages/labextension/src/model/assignment.ts
/**
 * Grader Extension API Schemas
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AssignmentSettings } from './assignmentSettings';

export interface Assignment {
  id?: number;
  name?: string;
  status?: Assignment.StatusEnum;
  points?: number;
  settings?: AssignmentSettings;
}
export namespace Assignment {
  export const StatusEnum = {
    Created: 'created',
    Pushed: 'pushed',
    Released: 'released',
    Complete: 'complete'
  } as const;
  export type StatusEnum = (typeof StatusEnum)[keyof typeof StatusEnum];
}